monad/requirements.txt
web3==6.11.3
python-dotenv==1.0.0
eth-account==0.10.0
requests==2.31.0
loguru==0.7.2
python-json-logger==2.0.7
colorama==0.4.6